example use and core functionality split (example working)
