+++ src/index.html
@@ -14,14 +14,6 @@
     <input type="file" id="fileinput" />
 
     <div id="output"></div>
-    <script src="dist/perf-cascade.js"></script>
-    <script>
-      //TODO: remove Dev/Test only - load test file
-      if (location.host.indexOf("127.0.0.1") === 0) {
-        //http://www.webpagetest.org/result/151226_X7_b43d35e592fab70e0ba012fe11a41020/
-        window["fetch"]("test-data/github.com.MODIFIED.151226_X7_b43d35e592fab70e0ba012fe11a41020.har")
-          .then(f => f.json().then(j => perfCascade.renderHar(j.log)))
-      }
-    </script>
+    <script src="dist/example-dist.js"></script>
   </body>
 </html>
